Modernize GH webhook test
